05_ecommerce_data_engineering_pipeline/fake_data_generator/airflow/requirements.txt
pandas
Faker
sqlalchemy
psycopg2-binary
pg8000
cloud-sql-python-connector
google-cloud-storage
